STM32 flash write SWD WAIT error workaround
This is fairly hacky...

Add retry logic so that we will retry mem_ap_read_atomic_u32()
upon WAIT errors.  These errors appear to be occurring due to the
flash write cycle hanging up the AHB bus.  This results in the status
check in cortex_m_poll() failing when it attempts to to read the
DCB_DHCSR register.  The WAIT error causes the target to go to an
unknown state.

This patch intercepts the WAIT as it propagates up the stack.  It
does it specifically in the context of the mem_ap_read_atomic_u32()
as that addresses the problem spot.  Due to the "dumbness" of the
FTDI device the SWD operations are queued up and written in a batch.
The status isn't observed until the whole batch is completed.  We
strive to isolate the operation in question so we can just retry
the read.

Overall this fix is narrow in scope and if the circumstances don't
align with the case observed, things will likely still break.